Use unsafe/reflect in serialization
Serialized representations are just little endian byte slices, which can be
obtained by referencing the data slice directly on certain platforms.
