+++ serialization_littleendian.go
@@ -1,0 +1,43 @@
+// +build 386 amd64,!appengine
+
+package roaring
+
+import (
+	"io"
+	"reflect"
+	"unsafe"
+)
+
+func (b *arrayContainer) writeTo(stream io.Writer) (int, error) {
+	buf := uint16SliceAsByteSlice(b.content)
+	return stream.Write(buf)
+}
+
+func (b *bitmapContainer) writeTo(stream io.Writer) (int, error) {
+	buf := uint64SliceAsByteSlice(b.bitmap)
+	return stream.Write(buf)
+}
+
+func uint64SliceAsByteSlice(slice []uint64) []byte {
+	// make a new slice header
+	header := *(*reflect.SliceHeader)(unsafe.Pointer(&slice))
+
+	// update its capacity and length
+	header.Len *= 8
+	header.Cap *= 8
+
+	// return it
+	return *(*[]byte)(unsafe.Pointer(&header))
+}
+
+func uint16SliceAsByteSlice(slice []uint16) []byte {
+	// make a new slice header
+	header := *(*reflect.SliceHeader)(unsafe.Pointer(&slice))
+
+	// update its capacity and length
+	header.Len *= 2
+	header.Cap *= 2
+
+	// return it
+	return *(*[]byte)(unsafe.Pointer(&header))
+}
